Add dib_role to the repo
image_building_dib is used to build images. It is
required to provide cim_image_info variable with
the images details as a list of dicts and all of
them will be built and placed in cim_build_dir
directory.

It is posible to define multiple images but they
will be built one by one. There is no reason to
build in paralel on a single build server as this
will not speedup the process.

Ready for use images are registered in
cim_ready_images variable as { label: file_path }